Update DekatronPC code with fine stats

+++ vhdl/run_tests.sh
@@ -5,11 +5,50 @@ trap cleanup SIGINT SIGTERM ERR EXIT
 
 script_dir=$(cd "$(dirname "${BASH_SOURCE[0]}")" &>/dev/null && pwd -P)
 
+png=0
+synt=0
+sim=0
+cov=0
 
 cleanup() {
     trap - SIGINT SIGTERM ERR EXIT
     exit
 
+}
+
+msg() {
+  echo >&2 -e "${1-}"
+}
+
+die() {
+  local msg=$1
+  local code=${2-1} # default exit status 1
+  msg "$msg"
+  exit "$code"
+}
+
+parse_params() {
+  # default values of variables set from params
+  flag=0
+  param=''
+
+  while :; do
+    case "${1-}" in
+    -h | --help) usage ;;
+    -v | --verbose) set -x ;;
+    -p | --png) no_png=1 ;;
+	-s | --synt) synt=1 ;;
+	-с | --coverage) cov=1 ;;
+	-t | --sim) sim=1 ;;
+    -?*) die "Unknown option: $1" ;;
+    *) break ;;
+    esac
+    shift
+  done
+
+  args=("$@")
+
+  return 0
 }
 
 emul() {
@@ -29,50 +68,80 @@ synt() {
 	echo 'tcl '${script_dir}'/synt_dpc.tcl '${files_path}' '${1}'' > ${1}.txt
 	cat ${1}.txt
 	yosys < ${1}.txt
-
 	cd ${current_dir}
 
 }
 
-DPCfiles=$(cat DPC.files)
+parse_params "$@"
 
-EmulFiles=$(cat Emul.files)
+if [ ${sim} -ne 0 ]; then
+	DPCfiles=$(cat DPC.files)
 
-emul Dekatron
+	EmulFiles=$(cat Emul.files)
 
-emul Counter
+	emul Dekatron
 
-#emul IpLine
+	emul Counter
 
-emul ApLine
+	#emul IpLine
 
-echo ${DPCfiles}
+	emul ApLine
 
-echo EmulFiles ${EmulFiles}
+	echo ${DPCfiles}
 
-verilator --top-module Emulator --lint-only -Wall ${EmulFiles} ${DPCfiles}
+	echo EmulFiles ${EmulFiles}
 
-verilator --top-module DekatronPC --lint-only  -Wall ${DPCfiles}
+	verilator --top-module Emulator --lint-only -Wall ${EmulFiles} ${DPCfiles}
 
-verilator -Wall --coverage --trace --top DekatronPC --cc ${DPCfiles} \
---timescale 100ns/100ps \
---exe DekatronPC/tests/DekatronPC.sv/DekatronPC_tb.cpp
+	verilator --top-module DekatronPC --lint-only  -Wall ${DPCfiles}
 
-make -j`nproc` -C obj_dir -f VDekatronPC.mk VDekatronPC
-./obj_dir/VDekatronPC
+	verilator -Wall --coverage --trace --top DekatronPC --cc ${DPCfiles} \
+	--timescale 100ns/100ps \
+	--exe DekatronPC/tests/DekatronPC.sv/DekatronPC_tb.cpp
 
-verilator_coverage -write-info logs/DPC.info logs/coverage_DPC.dat
+	make -j`nproc` -C obj_dir -f VDekatronPC.mk VDekatronPC
+	./obj_dir/VDekatronPC
 
-genhtml logs/DPC.info --output-directory coverage
+	if [ ! -d vcd ]; then
+		mkdir vcd
+	else
+		rm -f sch/*.vcd
+	fi
 
-synt DekatronPC
+	mv *.vcd vcd/
+	mv *UT  vcd/
+fi
 
-exit
-for file in $(ls *.dot); do
-		gvpr -f $script_dir/split $file
-	done
-	rm -f *.png
+if [ ${cov} -ne 0 ]; then
+	verilator_coverage -write-info logs/DPC.info logs/coverage_DPC.dat
+	genhtml logs/DPC.info --output-directory coverage
+fi
 
-	for file in $(ls *.dot); do 
-		echo $file; dot -Tpng $file -O
-	done
+if [ ${synt} -ne 0 ]; then
+	synt DekatronPC
+	python3 dpc_stat.py -j vtube.json -t DekatronPC -l vtube_cells.lib
+fi
+
+if [ ${png} -ne 0 ]; then
+	for file in $(ls *.dot); do
+			gvpr -f $script_dir/split $file
+		done
+		rm -f *.png
+
+		for file in $(ls *.dot); do 
+			echo $file; dot -Tpng $file -O
+			echo $file; dot -Tsvg $file -O
+		done
+
+	if [ ! -d sch ]; then
+		mkdir sch
+	else
+		rm -f sch/*.svg
+		rm -f sch/*.png
+		rm -f sch/*.dot
+	fi
+
+	mv *.svg sch/
+	mv *.png sch/
+	mv *.dot sch/
+fi
